Getting ready for v3.0.0 ● Moved source files to lib folder. ● Added index.js. ● Cleaned up documentation. ● Now using rollup as a sub-object for require.
