+++ lib/cbtCompile.js
