+++ lib/createEOTranslations.js
