+++ lib/getFileArrayFromGlob.js
