+++ lib/readFile.js
