+++ lib/readTranslations.js
