+++ lib/rollup.root.config.js
@@ -6,7 +6,7 @@ const path = require('path');
 const getFileArrayFromGlob = require('./getFileArrayFromGlob');
 const DIST_PATH_DEFAULT = 'dist/js';
 const PLUGIN_BUBLE = require('rollup-plugin-buble')({ transforms: { dangerousTaggedTemplateString: true } });
-const PLUGIN_UGLIFY = require('rollup-plugin-uglify-es')();
+const PLUGIN_UGLIFY = require('rollup-plugin-terser').terser({module: true});
 const ROOT = process.cwd();
 const BUILD_TYPES = {
   IIFE: 'IIFE', // IFFE output
