+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "component-build-tools",
-  "version": "2.0.1",
+  "version": "2.1.0",
   "description": "A set of build tools to simplify the creation of components",
-  "main": "rollup.root.config.js",
+  "main": "index.js",
   "author": "Michael G Collins <intervalia@gmail.com> http://www.intervalia.com/",
   "contributors": [
     "Michael G Collins <intervalia@gmail.com> http://www.intervalia.com/"
@@ -16,8 +16,9 @@
   },
   "dependencies": {
     "glob": "^7.1.2",
+    "rollup": "^0.60.7",
     "rollup-plugin-buble": "^0.19.2",
-    "rollup-plugin-uglify-es": "0.0.1"
+    "rollup-plugin-terser": "^1.0.1"
   },
   "devDependencies": {
     "chai": "^4.1.2",
